% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerschillendeFuncties.R
\name{cirkeldiagram}
\alias{cirkeldiagram}
\title{cirkeldiagram}
\usage{
cirkeldiagram(categorischevariabele, ...)
}
\arguments{
\item{categorischevariabele}{de variabele}

\item{...}{eventuele andere parameters}
}
\value{
cirkeldiagram
}
\description{
Functie maakt een cirkeldiagram van een categorische variabele, waarbij de categorieën in een legende verschijnen en de relatieve frequenties bij het cirkeldiagram zelf.
}
\examples{
cirkeldiagram(c('cat-a','cat-a','cat-b','cat-c','cat-a','cat-b'))
}
