% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerschillendeFuncties.R
\name{grafnorm}
\alias{grafnorm}
\title{grafnorm}
\usage{
grafnorm(mean = 0, sd = 1)
}
\arguments{
\item{mean}{gemiddelde van de normale verdeling (default = 0)}

\item{sd}{standaardafwijking van de normale verdeling (default = 1)}
}
\value{
grafiek
}
\description{
Functie tekent de grafiek van de dichtheidsfunctie bij een normale verdeling.
}
\examples{
grafnorm(0,1)
grafnorm(1,2)
}
