% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerschillendeFuncties.R
\name{kansbinom}
\alias{kansbinom}
\title{kansbinom}
\usage{
kansbinom(beginwaarde = 0, eindwaarde = NULL, n, p, ...)
}
\arguments{
\item{beginwaarde}{beginwaarde van het interval (default = 0)}

\item{eindwaarde}{eindwaarde van het interval (default = NULL)}

\item{n}{waarde van de parameter n (natuurlijk getal) horend bij de binomiale verdeling = aantal pogingen}

\item{p}{waarde van parameter p (0<p<1) horend bij de binomiale verdeling = kans op succes}

\item{...}{eventuele andere argumenten (bv. main)}
}
\value{
waarde van de kans en tekening van de oppervlakte onder de grafiek
}
\description{
Functie bepaalt bij een binomiale verdeling de kans en tekent de bijhorende oppervlakte.
}
\examples{
kansbinom(3,5,10,0.4)
kansbinom(eindwaarde=5,n=10,p=0.4)
kansbinom(beginwaarde=6,n=10,p=0.4,main='Laatste voorbeeld')
}
