% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerschillendeFuncties.R
\name{kansbinomiale}
\alias{kansbinomiale}
\title{kansbinomiale}
\usage{
kansbinomiale(kans, aantal, beginwaarde = 0, eindwaarde = NULL)
}
\arguments{
\item{kans}{waarde van parameter p (0<p<1) horend bij de binomiale verdeling}

\item{aantal}{waarde van de parameter n (natuurlijk getal) horend bij de normale verdeling}

\item{beginwaarde}{beginwaarde van het interval (default = 0)}

\item{eindwaarde}{eindwaarde van het interval (default = NULL)}
}
\value{
waarde van de kans en tekening van de oppervlakte onder de grafiek
}
\description{
Functie bepaalt bij een binomiale verdeling de kans en tekent de bijhorende oppervlakte.
}
\examples{
kansbinomiale(0.4,10,3,5)
}
