% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerschillendeFuncties.R
\name{kansnorm}
\alias{kansnorm}
\title{kansnorm}
\usage{
kansnorm(beginwaarde = -Inf, eindwaarde = +Inf, mean = 0, sd = 1, ...)
}
\arguments{
\item{beginwaarde}{beginwaarde van het interval (default = -oneindig)}

\item{eindwaarde}{eindwaarde van het interval (default = + oneindig)}

\item{mean}{gemiddelde van de normale verdeling (default = 0)}

\item{sd}{standaardafwijking van de normale verdeling (default = 1)}

\item{...}{eventuele andere argumenten (bv. main)}
}
\value{
waarde van de kans en tekening van oppervlakte onder dichtheidsfunctie
}
\description{
Functie bepaalt bij een normale verdeling de kans en tekent de bijhorende oppervlakte.
}
\examples{
kansnorm(beginwaarde=2)
kansnorm(beginwaarde=-1,eindwaarde=1)
kansnorm(beginwaarde=0,mean=1,sd=2,main="Laatste voorbeeld")
}
