% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerschillendeFuncties.R
\name{kansnormaleverdeling}
\alias{kansnormaleverdeling}
\title{kansnormaleverdeling}
\usage{
kansnormaleverdeling(
  beginwaarde = -Inf,
  eindwaarde = +Inf,
  gemiddelde = 0,
  standaardafwijking = 1
)
}
\arguments{
\item{beginwaarde}{beginwaarde van het interval (default = -oneindig)}

\item{eindwaarde}{eindwaarde van het interval (default = + oneindig)}

\item{gemiddelde}{gemiddelde van de normale verdeling (default = 0)}

\item{standaardafwijking}{standaardafwijking van de normale verdeling (default = 1)}
}
\value{
waarde van de kans en tekening van oppervlakte onder dichtheidsfunctie
}
\description{
Functie bepaalt bij een normale verdeling de kans en tekent de bijhorende oppervlakte.
}
\examples{
kansnormaleverdeling(beginwaarde=2)
}
