% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerschillendeFuncties.R
\name{modus}
\alias{modus}
\title{Modus}
\usage{
modus(x)
}
\arguments{
\item{x}{een dataset}
}
\value{
de modus of modi van x
}
\description{
Functie bepaalt de centrummaat modus van een dataset.
Indien er verschillende waardes zijn die het meest voorkomen, dan worden deze allemaal bepaald.
}
\examples{
modus(c(1,1,2,2,3))
modus(c('a','u','t','o','b','u','s'))
}
