% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerschillendeFuncties.R
\name{normaldichtheidsfunctie}
\alias{normaldichtheidsfunctie}
\title{normaldichtheidsfunctie}
\usage{
normaldichtheidsfunctie(gemiddelde = 0, standaardafwijking = 1)
}
\arguments{
\item{gemiddelde}{gemiddelde van de normale verdeling (default = 0)}

\item{standaardafwijking}{standaardafwijking van de normale verdeling (default = 1)}
}
\value{
grafiek van de dichtheidsfunctie
}
\description{
Functie tekent de dichtheidsfunctie van een normale verdeling.
}
\examples{
normaldichtheidsfunctie(0,1)
}
